backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
pgvector
pypdf
langchain
langchain_openai
langchain_text_splitters
langchain-community
